% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requal_module.R
\name{mod_server}
\alias{mod_server}
\title{rql.textimport Server Functions}
\usage{
mod_server(id, api)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}

\item{api}{Connection object for Requal database.}
}
\description{
rql.textimport Server Functions
}
